Move Vehicile Repair/Refuel harmony patches to specific directories
